requirements.txt
# Core LangChain packages - pinned to compatible versions
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.1

# Vector store and embeddings
chromadb>=0.4.22

# PDF processing
PyMuPDF>=1.23.0
Pillow>=10.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# HTTP requests for Ollama API
requests>=2.31.0